+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add dialog warning user when trying to skip external download mods from CurseForge
 - Add logging for Java version and install method on launch if analytics enabled
 - Log warnings when useSystemGlfw/useSystemOpenAl are enabled when launching
+- Add popup when Minecraft crashes due to needing Java 17
 
 ### Fixes
 - Remove old OmitStackTraceInFastThrow JVM arg